Finished a refactor for the fast activation saving.
