+++ torchlens/__init__.py
@@ -1,4 +1,4 @@
 """ Top level package: make the user-facing functions top-level, rest accessed as submodules.
 """
-from torchlens.user_funcs import log_forward_pass, get_model_metadata, show_model_graph, \
+from torchlens.user_funcs import log_forward_pass, show_model_graph, \
     validate_saved_activations, validate_batch_of_models_and_inputs
